fix: add right math classes

+++ src/main/java/com/cgvsu/math/Vector3f.java
@@ -23,7 +23,7 @@ public class Vector3f {
         return y;
     }
 
-    public double getZ() {
+    public float getZ() {
         return z;
     }
